### 2020-10-23 #2773 @whedon
[PRE REVIEW]: pyEQUIB: Python Package for Plasma Diagnostics and Abundance Analysis
**Submitting author:** @danehkar (<a href="http://orcid.org/0000-0003-4552-5997">Ashkbiz Danehkar</a>)
**Repository:** <a href="https://github.com/equib/pyEQUIB" target ="_blank">https://github.com/equib/pyEQUIB</a>
**Version:** v0.3.3
