+++ paper/paper.md
@@ -40,7 +40,7 @@ The program `EQUIB` computes atomic level populations and emissivities for multi
 
 - the _reddening unit_ that contains API functions for the extinction calculation and reddening correction. This unit was developed based on the reddening IRAF scripts included in the Space Telescope Science Data Analysis System, former `STSDAS` IRAF package [@Bushouse:1994; @Shaw:1994]. These API functions can be used to calculate the interstellar extinction from the observed Balmer decrements for different reddening laws, and perform the reddening correction on the observed fluxes of nebular emission lines. 
 
-`pyEQUIB` can be simply used by astronomers, who are familiar with the high-level, general-purpose programming language Python. The previous IDL version `proEQUIB` [@Danehkar:2018b] has been used for studies of gaseous nebulae [@Danehkar:2016; @Danehkar:2018a]. This package requires the Python packages `NumPy` [@Walt:2011; @Harris:2020], `SciPy` [@Virtanen:2020], and `AtomNeb` [@Danehkar:2020]. The API functions of this package can be used to analyze emission-line spectra from planetary nebulae and H II regions, as well as extragalactic sources [see e.g. @Danehkar:2014; @Danehkar:2014b; @Danehkar:2016]. This package is released under the GNU General Public License. The source code is publicly available on the GitHub platform. The latest version of this package can be installed directly from its repository on the GitHub, and its stable version from the Python Package Index (PyPi) via ``pip install pyequib`` or alternatively from the Anaconda Python package distributor via ``conda install -c conda-forge pyequib``. The online documentation, tutorials and examples are available on its GitHub page (https://equib.github.io/pyEQUIB/) and its Read the Docs documentation page (https://pyequib.readthedocs.io/).
+`pyEQUIB` can simply be used by astronomers, who are familiar with the high-level, general-purpose programming language Python. The previous IDL version `proEQUIB` [@Danehkar:2018b] has been used for studies of gaseous nebulae [@Danehkar:2016; @Danehkar:2018a]. This package requires the Python packages `NumPy` [@Walt:2011; @Harris:2020], `SciPy` [@Virtanen:2020], and `AtomNeb` [@Danehkar:2020]. The API functions of this package can be used to analyze emission-line spectra from planetary nebulae and H II regions, as well as extragalactic sources [see e.g. @Danehkar:2014; @Danehkar:2014b; @Danehkar:2016]. This package is released under the GNU General Public License. The source code is publicly available on the GitHub platform. The latest version of this package can be installed directly from its repository on the GitHub, and its stable version from the Python Package Index (PyPi) via ``pip install pyequib`` or alternatively from the Anaconda Python package distributor via ``conda install -c conda-forge pyequib``. The online documentation, tutorials and examples are available on its GitHub page (https://equib.github.io/pyEQUIB/) and its Read the Docs documentation page (https://pyequib.readthedocs.io/).
 
 # Acknowledgements
 
