### 2020-10-28 #2773 @arfon
Hi @danehkar, many thanks for your submission to JOSS.

After discussing with the JOSS editorial team, we've determined that https://github.com/equib/pyEQUIB is out of scope for JOSS as it is a direct port of your earlier package (https://github.com/equib/proEQUIB) and therefore does not reach our threshold for [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort).

While we can't support a pulication in JOSS related to this work, we would allow an update (addendum) to [your paper](https://joss.theoj.org/papers/10.21105/joss.00899) on `proEQUIB` noting that there is now a Python port of this package.

Please let me know if this is of interest to you and if so, I'll be happy to work with you on an update to your earlier article.
