### 2020-10-28 #2773 @danehkar
> After discussing with the JOSS editorial team, we've determined that https://github.com/equib/pyEQUIB is out of scope for JOSS as it is a direct port of your earlier package (https://github.com/equib/proEQUIB) and therefore does not reach our threshold for [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort).

@arfon, actually, a direct port is not a correct description. There is a possibility to port IDL into Python. However, the new python package is not a port of IDL version, it is purely a new implementation in Python language and translation of IDL version into Python code. You can see the Python codes here:

https://github.com/equib/pyEQUIB/blob/master/pyequib/pyequib.py

IDL verison depends on the IDL Astronomy User's Library (https://idlastro.gsfc.nasa.gov/homepage.html), while the Python version depends on AstroPy (https://www.astropy.org/). Atomic data is stored in the structure in IDL version, while atomic data is stored using disctionary and pandas library in Python. Interpolation in Python is done using some functions from SciPy.

> While we can't support a pulication in JOSS related to this work, we would allow an update (addendum) to your paper on proEQUIB noting that there is now a Python port of this package.

I think that the correct phrase is a Python implementation (not a Python port).of this package.
