### 2020-10-28 #2773 @danehkar
@arfon For porting IDL package into Python, we should use [idlwrap](https://pypi.org/project/idlwrap/), https://r4lv.github.io/idlwrap/Porting%20IDL%20to%20Python.html, which is not the case her. Here we have a translation and partly implementation of this package purely in the Python language without the use of IDL language . I do not use any library such as [idlwrap](https://pypi.org/project/idlwrap/) to port IDL codes into Python. I rewrote and translated all functions purely into Python, and this package do not depend on its IDL version.
