### 2020-10-28 #2773 @danehkar
@arfon I am happy to update the paper and mention its Python implementation. Can we also revise the title of my published paper? For example, to "proEQUIB/pyEQUIB: IDL/Python Package for Plasma Diagnostics and Abundance Analysis"
