### 2020-10-28 #2773 @arfon
> @arfon I am happy to update the published paper and mention its Python implementation. Can we also revise the title of my published paper? For example, to "proEQUIB/pyEQUIB: IDL/Python Package for Plasma Diagnostics and Abundance Analysis"

I don't think we can modify the old paper sorry - this is considered [bad practice](https://www.stm-assoc.org/2017_09_05_STM_Guide_Preserving_the_Record_of_Science_5_September_2017.pdf). The addendum would be a separate (very short) note describing pyEQUIB. The addendum will be a citable object in itself (i.e. will have its own DOI) and we'll make sure it links back to the proEQUIB paper.

We could try and include `pyEQUIB` in the title of the addendum perhaps e.g.:

`pyEQUIB, an addendum to proEQUIB: IDL Library for Plasma Diagnostics and Abundance Analysis`

Or something like that?
