### 2020-10-28 #2773 @danehkar
> I don't think we can modify the old paper sorry - this is considered [bad practice](https://www.stm-assoc.org/2017_09_05_STM_Guide_Preserving_the_Record_of_Science_5_September_2017.pdf). The addendum would be a separate (very short) note describing pyEQUIB. The addendum will be a citable object in itself (i.e. will have its own DOI) and we'll make sure it links back to the proEQUIB paper.

You probably need to include the following text in addendum:

The Python package `pyEQUIB` requires the Python packages `NumPy` [@Walt:2011; @Harris:2020], `SciPy` [@Virtanen:2020], and `AtomNeb`. The latest version of this package can be installed directly from its repository on the GitHub using ``python setup.py install``, and its stable version from the Python Package Index (PyPi) via ``pip install pyequib`` or alternatively from the Anaconda Python package distributor via ``conda install -c conda-forge pyequib``. The online documentation, tutorials and examples are available on its GitHub page (https://equib.github.io/pyEQUIB/) and its Read the Docs documentation page (https://pyequib.readthedocs.io/).

@ARTICLE{Harris:2020,
       author = {{Harris}, C. R. and {Jarrod Millman}, K. and
         {van der Walt}, S. J. and {Gommers}, R. and
         {Virtanen}, P. and {Cournapeau}, D. and {Wieser}, E. and
         {Taylor}, Julian and {Berg}, Sebastian and {Smith}, Nathaniel J. and
         {Kern}, Robert and {Picus}, Matti and {Hoyer}, Stephan and
         {van Kerkwijk}, Marten H. and {Brett}, Matthew and {Haldane}, Allan and
         {Fern{\'a}ndez del R{\'\i}o}, Jaime and {Wiebe}, Mark and
         {Peterson}, Pearu and {G{\'e}rard-Marchant}, Pierre and
         {Sheppard}, Kevin and {Reddy}, Tyler and {Weckesser}, Warren and
         {Abbasi}, Hameer and {Gohlke}, Christoph and {Oliphant}, Travis E.},
        title = "{Array Programming with NumPy}",
      journal = {Nature},
         year = 2020,
       volume = {585},
        pages = {357},
          doi = {10.1038/s41586-020-2649-2},
archivePrefix = {arXiv},
       eprint = {2006.10256},
 primaryClass = {cs.MS},
       adsurl = {https://ui.adsabs.harvard.edu/abs/2020arXiv200610256H},
      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
}

@ARTICLE{Virtanen:2020,
       author = {{Virtanen}, P. and {Gommers}, R. and {Oliphant}, T. E. and
         {Haberland}, M. and {Reddy}, Tyler and {Cournapeau}, David and
         {Burovski}, Evgeni and {Peterson}, Pearu and {Weckesser}, Warren and
         {Bright}, Jonathan and {van der Walt}, St{\'e}fan J. and
         {Brett}, Matthew and {Wilson}, Joshua and {Millman}, K. Jarrod and
         {Mayorov}, Nikolay and {Nelson}, Andrew R.~J. and {Jones}, Eric and
         {Kern}, Robert and {Larson}, Eric and {Carey}, C.~J. and
         {Polat}, {\.I}lhan and {Feng}, Yu and {Moore}, Eric W. and {Vand
        erPlas}, Jake and {Laxalde}, Denis and {Perktold}, Josef and
         {Cimrman}, Robert and {Henriksen}, Ian and {Quintero}, E.~A. and
         {Harris}, Charles R. and {Archibald}, Anne M. and
         {Ribeiro}, Ant{\^o}nio H. and {Pedregosa}, Fabian and
         {van Mulbregt}, Paul and {SciPy 1. 0 Contributors}},
        title = "{SciPy 1.0: fundamental algorithms for scientific computing in Python}",
      journal = {Nature Methods},
         year = 2020,
        month = feb,
       volume = {17},
        pages = {261-272},
          doi = {10.1038/s41592-019-0686-2},
archivePrefix = {arXiv},
       eprint = {1907.10121},
 primaryClass = {cs.MS},
       adsurl = {https://ui.adsabs.harvard.edu/abs/2020NatMe..17..261V},
      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
}

@ARTICLE{Walt:2011,
       author = {{van der Walt}, S. and {Colbert}, S. C. and
         {Varoquaux}, G.},
        title = "{The NumPy Array: A Structure for Efficient Numerical Computation}",
      journal = {Computing in Science and Engineering},
         year = 2011,
       volume = {13},
       number = {2},
        pages = {22-30},
          doi = {10.1109/MCSE.2011.37},
archivePrefix = {arXiv},
       eprint = {1102.1523},
 primaryClass = {cs.MS},
       adsurl = {https://ui.adsabs.harvard.edu/abs/2011CSE....13b..22V},
      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
}
