### 2020-10-28 #2773 @danehkar
> We could try and include `pyEQUIB` in the title of the addendum perhaps e.g.:
>
> `pyEQUIB, an addendum to proEQUIB: IDL Library for Plasma Diagnostics and Abundance Analysis`
>
> Or something like that?

@arfon it is OK.
