### 2020-10-28 #2773 @danehkar
@arfon

It is better to be `pyEQUIB Python Package, an addendum to AtomNeb: IDL Library for Atomic Data of Ionized Nebulae`
