### 2020-10-29 #2798 @arfon
@danehkar - please adapt your `paper.md` to follow the same format as I described over here: https://github.com/openjournals/joss-reviews/issues/2772#issuecomment-718227083

Once you've done that, I can work on getting the addendum published.
