+++ paper/paper.bib
@@ -1,77 +1,4 @@
 
-@ARTICLE{Danehkar:2013,
-   author = {{Danehkar}, A. and {Parker}, Q.~A. and {Ercolano}, B.},
-    title = "{Observations and three-dimensional ionization structure of the planetary nebula SuWt 2}",
-  journal = {Monthly Notices of the Royal Astronomical Society},
-archivePrefix = "arXiv",
-   eprint = {1307.2974},
- primaryClass = "astro-ph.SR",
-     year = 2013,
-   volume = 434,
-    pages = {1513-1530},
-      doi = {10.1093/mnras/stt1116},
-   adsurl = {http://adsabs.harvard.edu/abs/2013MNRAS.434.1513D},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@ARTICLE{Danehkar:2014,
-   author = {{Danehkar}, A. and {Todt}, H. and {Ercolano}, B. and {Kniazev}, A.~Y.
-	},
-    title = "{Observations and three-dimensional photoionization modelling of the Wolf-Rayet planetary nebula Abell 48}",
-  journal = {Monthly Notices of the Royal Astronomical Society},
-archivePrefix = "arXiv",
-   eprint = {1403.0567},
- primaryClass = "astro-ph.SR",
-     year = 2014,
-   volume = 439,
-    pages = {3605-3615},
-      doi = {10.1093/mnras/stu203},
-   adsurl = {http://adsabs.harvard.edu/abs/2014MNRAS.439.3605D},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@PHDTHESIS{Danehkar:2014b,
-   author = {{Danehkar}, A.},
-    title = "{Evolution of Planetary Nebulae with WR-type Central Stars}",
-   school = {Macquarie University, Australia},
-     year = 2014,
-      doi = {10.5281/zenodo.47794},
-   adsurl = {http://adsabs.harvard.edu/abs/2014PhDT........76D},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@ARTICLE{Danehkar:2016,
-   author = {{Danehkar}, A. and {Parker}, Q.~A. and {Steffen}, W.},
-    title = "{Fast, Low-ionization Emission Regions of the Planetary Nebula M2-42}",
-  journal = {The Astronomical Journal},
-archivePrefix = "arXiv",
-   eprint = {1601.01702},
- primaryClass = "astro-ph.SR",
-     year = 2016,
-   volume = 151,
-      eid = {38},
-    pages = {38},
-      doi = {10.3847/0004-6256/151/2/38},
-   adsurl = {http://adsabs.harvard.edu/abs/2016AJ....151...38D},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@ARTICLE{Danehkar:2018a,
-   author = {{Danehkar}, A.},
-    title = "{Bi-Abundance Ionisation Structure of the Wolf-Rayet Planetary Nebula PB 8}",
-  journal = {Publications of the Astronomical Society of Australia},
-archivePrefix = "arXiv",
-   eprint = {1801.00892},
- primaryClass = "astro-ph.SR",
-     year = 2018,
-   volume = 35,
-      eid = {e005},
-    pages = {e005},
-      doi = {10.1017/pasa.2018.1},
-   adsurl = {http://adsabs.harvard.edu/abs/2018PASA...35....5D},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
 @ARTICLE{Danehkar:2018b,
    author = {{Danehkar}, A.},
[…]
