+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: "pyEQUIB: Python Package for Plasma Diagnostics and Abundance Analysis"
+title: "pyEQUIB Python Package, an addendum to proEQUIB: IDL Library for Plasma Diagnostics and Abundance Analysis"
 tags:
   - python
   - astrophysics
@@ -21,26 +21,11 @@ date: 20 October 2020
 bibliography: paper.bib
 ---
 
-# Introduction
+# Addendum
 
-The interpretation of emission-line spectra from ionized gases are crucial in studying planetary nebulae and H II regions in our Galaxy and other galaxies. Physical properties of the interstellar medium such as the electron temperature and concentration can be determined from nebular emission lines. Moreover, chemical elements can be probed using the collisionally excited lines (CEL) and recombination lines (RL) emitted from ionized gases in the interstellar medium [see e.g. @Danehkar:2013; @Danehkar:2016; @Danehkar:2018a]. The interstellar extinction can also be derived from the observed Balmer decrements. 
+`pyEQUIB` is a pure Python open-source package containing several application programming interface (API) functions that can be employed for plasma diagnostics and abundance analysis of nebular emission lines. This package is a Python implementation of the IDL library `proEQUIB` [@Danehkar:2018b] that is coupled to the IDL library `AtomNeb` [@Danehkar:2019]. The collisional excitation and recombination units of this package need to have the energy levels, collision strengths, transition probabilities, and recombination coefficients, which can be retrieved from the Python package `AtomNeb` for _Atomic Data of Ionized Nebulae_ [@Danehkar:2020]. The API functions of this package can be used to deduce the electron temperature, electron concentration, chemical elements from CELs and Rls, and the interstellar extinction from the Balmer decrements emitted from ionized gaseous nebulae. This package can simply be used by astronomers, who are familiar with the high-level, general-purpose programming language Python.
 
-# Statement of Need
-
-The spectral analysis of nebular emission lines with the Python programming language requires a package that can calculate emissivities for collisional excitation and recombination processes with the given electron temperature and density, and estimate the  reddening factor for the given interstellar extinction law and physical condition. The previous package `proEQUIB` [@Danehkar:2018b] requires the Interactive Data Language (IDL) compiler. It was necessary to make a similar package purely developed in Python that can be widely used by astronomers and researchers, who prefer to conduct the data analysis of nebular spectra with the high-level programming language Python that is freely available for all operating systems.
-
-# Description
-
-`pyEQUIB` is a pure Python open-source package containing several application programming interface (API) functions that can be employed for plasma diagnostics and abundance analysis of nebular emission lines. This package is a Python implementation of the IDL library `proEQUIB` [@Danehkar:2018b] that is coupled to the IDL library `AtomNeb` [@Danehkar:2019]. The collisional excitation and recombination units of this package need to have the energy levels, collision strengths, transition probabilities, and recombination coefficients, which can be retrieved from the Python package `AtomNeb` for _Atomic Data of Ionized Nebulae_ [@Danehkar:2020]. The API functions of this package can be used to deduce the electron temperature, electron concentration, chemical elements from CELs and Rls, and the interstellar extinction from the Balmer decrements emitted from ionized gaseous nebulae. This package has three units, namely:
-
-- the _collisional excitation unit_ that contains API functions for the interpretation of CELs. This unit was initially developed based on the algorithm of the FORTRAN program `EQUIB` [@Howarth:1981; @Howarth:2016], which was also incorporated into the nebular empirical analysis tool `NEAT` [@Wesson:2012]. 
[…]
