### 2020-11-23 #2798 @danehkar
@arfon I made a new release (v0.3.3) in its GitHub repository ([https://github.com/equib/pyEQUIB](https://github.com/equib/pyEQUIB)) and also an archive of the software in Zenodo service:

doi:[10.5281/zenodo.4287576](https://doi.org/10.5281/zenodo.4287576)
